Safari-611.1.10/LayoutTests/fast/dom/title-text-property-2-expected.txt

CONSOLE MESSAGE: Setting document.title to TITLE1
TITLE CHANGED: 'TITLE1'
CONSOLE MESSAGE: Setting title element's text to TITLE2
TITLE CHANGED: 'TITLE2'
CONSOLE MESSAGE: Should not set title
CONSOLE MESSAGE: Should reset title to the empty string
TITLE CHANGED: ''
